Add authentication check when a PUT, DELETE RestfulAPI
When registering or deleting answers, it checks whether the administrator is authorized through Headers.

+++ src/server/routes/api.ts
@@ -1,5 +1,6 @@
 import firebase from '../firebase';
 import {express, logger} from '../app';
+import config from '../config';
 
 const router = express.Router();
 
@@ -21,18 +22,30 @@ router.get('/', (req, res) => {
 
 router.put('/:id/:reply', (req, res) => {
     const {id, reply} = req.params;
+    const admin = req.headers['x-askq-auth'];
 
-    logger.info(`PUT /question/${id}/${reply}`);
-    firebase.updateQuestion(id, reply);
-    res.sendStatus(200).end();
+    if(admin !== undefined && admin === config.adminUid) {
+        logger.info(`PUT /question/${id}/${reply}`);
+        firebase.updateQuestion(id, reply);
+        res.sendStatus(200).end();
+        return;
+    }
+
+    res.sendStatus(401).end();
 });
 
 router.delete('/:id', (req, res) => {
     const {id} = req.params;
+    const admin = req.headers['x-askq-auth'];
 
-    logger.info(`DELETE /question/${id}`);
-    firebase.deleteQuestion(id);
-    res.sendStatus(200).end();
+    if(admin !== undefined && admin === config.adminUid) {
+        logger.info(`DELETE /question/${id}`);
+        firebase.deleteQuestion(id);
+        res.sendStatus(200).end();
+        return;
+    }
+
+    res.sendStatus(401).end();
 });
 
 export default {
